Refactor Tablet for cloud object storage
The original Tablet is abstracted into two subclasses.
1. LocalTablet is based on local disk storage and replicas are managed by StarRocks.
2. StarOSTablet is based on cloud object storage. Data replicas are managed by object storage
and compute replicas are managed by StarOS through Shard.

Update the Gson persistence framework.
1. Support deserializing LocalTablet from old Tablet serialized data when upgrading.
2. Support deserializing old Tablet from LocalTablet serialized data when rolling back.
