+++ fe/fe-core/src/main/java/com/starrocks/alter/SchemaChangeJob.java
@@ -34,6 +34,7 @@ import com.starrocks.catalog.Catalog;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.Database;
 import com.starrocks.catalog.KeysType;
+import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.MaterializedIndex.IndexExtState;
 import com.starrocks.catalog.MaterializedIndex.IndexState;
@@ -339,7 +340,7 @@ public class SchemaChangeJob extends AlterJob {
                 long partitionId = partition.getId();
                 for (MaterializedIndex index : partition.getMaterializedIndices(IndexExtState.ALL)) {
                     for (Tablet tablet : index.getTablets()) {
-                        List<Replica> replicas = tablet.getReplicas();
+                        List<Replica> replicas = ((LocalTablet) tablet).getReplicas();
                         for (Replica replica : replicas) {
                             long backendId = replica.getBackendId();
                             ClearAlterTask clearAlterTask = new ClearAlterTask(backendId, dbId, tableId,
@@ -430,7 +431,7 @@ public class SchemaChangeJob extends AlterJob {
                         for (Tablet tablet : alterIndex.getTablets()) {
                             long tabletId = tablet.getId();
                             short replicaSendNum = 0;
-                            for (Replica replica : tablet.getReplicas()) {
+                            for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                                 if (replica.getState() != ReplicaState.SCHEMA_CHANGE) {
                                     // for now, all replica should be in SCHEMA_CHANGE,
                                     // because we don't allow tablet repair and balance during schema change.
@@ -517,7 +518,7 @@ public class SchemaChangeJob extends AlterJob {
                     }
                     for (Tablet tablet : index.getTablets()) {
                         long tabletId = tablet.getId();
-                        for (Replica replica : tablet.getReplicas()) {
+                        for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                             if (replica.getState() == ReplicaState.CLONE
                                     || replica.getState() == ReplicaState.DECOMMISSION
                                     || replica.getState() == ReplicaState.NORMAL) {
@@ -608,7 +609,7 @@ public class SchemaChangeJob extends AlterJob {
             Preconditions.checkState(materializedIndex.getState() == IndexState.SCHEMA_CHANGE);
 
             Preconditions.checkArgument(finishTabletInfo.getTablet_id() == tabletId);
-            Tablet tablet = materializedIndex.getTablet(tabletId);
+            LocalTablet tablet = (LocalTablet) materializedIndex.getTablet(tabletId);
             if (tablet == null) {
                 throw new MetaNotFoundException("Cannot find tablet[" + tabletId + "]");
             }
@@ -690,7 +691,7 @@ public class SchemaChangeJob extends AlterJob {
                         }
 
                         for (Tablet tablet : materializedIndex.getTablets()) {
-                            List<Replica> replicas = tablet.getReplicas();
+                            List<Replica> replicas = ((LocalTablet) tablet).getReplicas();
                             List<Replica> errorReplicas = Lists.newArrayList();
                             int healthNum = replicas.size();
                             for (Replica replica : replicas) {
@@ -762,7 +763,7 @@ public class SchemaChangeJob extends AlterJob {
                         if (!this.partitionIdToFinishedIndexIds.containsKey(partitionId)
                                 || !this.partitionIdToFinishedIndexIds.get(partitionId).contains(indexId)) {
                             for (Tablet tablet : materializedIndex.getTablets()) {
-                                for (Replica replica : tablet.getReplicas()) {
+                                for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                                     AgentTaskQueue.removeTask(replica.getBackendId(), TTaskType.SCHEMA_CHANGE,
                                             tablet.getId());
                                 }
@@ -800,7 +801,7 @@ public class SchemaChangeJob extends AlterJob {
                         Preconditions.checkState(materializedIndex.getState() == IndexState.SCHEMA_CHANGE);
                         for (Tablet tablet : materializedIndex.getTablets()) {
                             long tabletId = tablet.getId();
-                            for (Replica replica : tablet.getReplicas()) {
+                            for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                                 if (replica.getState() != ReplicaState.SCHEMA_CHANGE) {
                                     // all replicas should be in state SCHEMA_CHANGE
                                     cancelMsg = String.format(
@@ -919,7 +920,7 @@ public class SchemaChangeJob extends AlterJob {
                     MaterializedIndex index = partition.getIndex(entry.getKey());
                     // set state to SCHEMA_CHANGE
                     for (Tablet tablet : index.getTablets()) {
-                        for (Replica replica : tablet.getReplicas()) {
+                        for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                             if (replica.getState() == ReplicaState.CLONE
                                     || replica.getState() == ReplicaState.DECOMMISSION) {
                                 // add log here, because there should no more CLONE replica when processing alter jobs.
@@ -962,7 +963,7 @@ public class SchemaChangeJob extends AlterJob {
                 for (Map.Entry<Long, Integer> entry : changedIndexIdToSchemaHash.entrySet()) {
                     MaterializedIndex index = partition.getIndex(entry.getKey());
                     for (Tablet tablet : index.getTablets()) {
-                        for (Replica replica : tablet.getReplicas()) {
+                        for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                             replica.setState(ReplicaState.NORMAL);
                         }
                     }
@@ -985,7 +986,7 @@ public class SchemaChangeJob extends AlterJob {
                         if (schemaHash == -1) {
                             schemaHash = getSchemaHashByIndexId(info.getIndexId());
                         }
-                        Tablet tablet = mIndex.getTablet(info.getTabletId());
+                        LocalTablet tablet = (LocalTablet) mIndex.getTablet(info.getTabletId());
                         if (info.getOpType() == ReplicaOperationType.SCHEMA_CHANGE) {
                             Replica replica = tablet.getReplicaByBackendId(info.getBackendId());
                             replica.updateVersionInfo(info.getVersion(),
@@ -1062,7 +1063,7 @@ public class SchemaChangeJob extends AlterJob {
                         continue;
                     }
                     for (Tablet tablet : index.getTablets()) {
-                        for (Replica replica : tablet.getReplicas()) {
+                        for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                             if (replica.getState() == ReplicaState.CLONE
                                     || replica.getState() == ReplicaState.DECOMMISSION
                                     || replica.getState() == ReplicaState.NORMAL) {
