+++ fe/fe-core/src/main/java/com/starrocks/alter/SchemaChangeJobV2.java
@@ -35,6 +35,7 @@ import com.starrocks.catalog.Column;
 import com.starrocks.catalog.Database;
 import com.starrocks.catalog.Index;
 import com.starrocks.catalog.KeysType;
+import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.MaterializedIndex.IndexState;
 import com.starrocks.catalog.OlapTable;
@@ -222,7 +223,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
         int totalReplicaNum = 0;
         for (MaterializedIndex shadowIdx : partitionIndexMap.values()) {
             for (Tablet tablet : shadowIdx.getTablets()) {
-                totalReplicaNum += tablet.getReplicas().size();
+                totalReplicaNum += ((LocalTablet) tablet).getReplicas().size();
             }
         }
         MarkedCountDownLatch<Long, Long> countDownLatch = new MarkedCountDownLatch<>(totalReplicaNum);
@@ -268,7 +269,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
 
                     for (Tablet shadowTablet : shadowIdx.getTablets()) {
                         long shadowTabletId = shadowTablet.getId();
-                        List<Replica> shadowReplicas = shadowTablet.getReplicas();
+                        List<Replica> shadowReplicas = ((LocalTablet) shadowTablet).getReplicas();
                         for (Replica shadowReplica : shadowReplicas) {
                             long backendId = shadowReplica.getBackendId();
                             countDownLatch.addMark(backendId, shadowTabletId);
@@ -429,7 +430,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
                     for (Tablet shadowTablet : shadowIdx.getTablets()) {
                         long shadowTabletId = shadowTablet.getId();
                         long originTabletId = partitionIndexTabletMap.get(partitionId, shadowIdxId).get(shadowTabletId);
-                        for (Replica shadowReplica : shadowTablet.getReplicas()) {
+                        for (Replica shadowReplica : ((LocalTablet) shadowTablet).getReplicas()) {
                             AlterReplicaTask rollupTask = new AlterReplicaTask(
                                     shadowReplica.getBackendId(), dbId, tableId, partitionId,
                                     shadowIdxId, originIdxId,
@@ -519,7 +520,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
                     MaterializedIndex shadowIdx = entry.getValue();
 
                     for (Tablet shadowTablet : shadowIdx.getTablets()) {
-                        List<Replica> replicas = shadowTablet.getReplicas();
+                        List<Replica> replicas = ((LocalTablet) shadowTablet).getReplicas();
                         int healthyReplicaNum = 0;
                         for (Replica replica : replicas) {
                             if (replica.getLastFailedVersion() < 0
@@ -575,7 +576,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
 
                 // set the replica state from ReplicaState.ALTER to ReplicaState.NORMAL since the schema change is done.
                 for (Tablet tablet : shadowIdx.getTablets()) {
-                    for (Replica replica : tablet.getReplicas()) {
+                    for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                         replica.setState(ReplicaState.NORMAL);
                     }
                 }
@@ -721,7 +722,7 @@ public class SchemaChangeJobV2 extends AlterJobV2 {
 
                 for (Tablet shadownTablet : shadowIndex.getTablets()) {
                     invertedIndex.addTablet(shadownTablet.getId(), shadowTabletMeta);
-                    for (Replica shadowReplica : shadownTablet.getReplicas()) {
+                    for (Replica shadowReplica : ((LocalTablet) shadownTablet).getReplicas()) {
                         invertedIndex.addReplica(shadownTablet.getId(), shadowReplica);
                     }
                 }
