+++ fe/fe-core/src/main/java/com/starrocks/planner/OlapTableSink.java
@@ -42,6 +42,7 @@ import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.PartitionKey;
 import com.starrocks.catalog.PartitionType;
 import com.starrocks.catalog.RangePartitionInfo;
+import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.Tablet;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.DdlException;
@@ -305,11 +306,12 @@ public class OlapTableSink extends DataSink {
                 // we should ensure the replica backend is alive
                 // otherwise, there will be a 'unknown node id, id=xxx' error for stream load
                 for (Tablet tablet : index.getTablets()) {
-                    Multimap<Long, Long> bePathsMap = tablet.getNormalReplicaBackendPathMap(table.getClusterId());
+                    LocalTablet localTablet = (LocalTablet) tablet;
+                    Multimap<Long, Long> bePathsMap = localTablet.getNormalReplicaBackendPathMap(table.getClusterId());
                     if (bePathsMap.keySet().size() < quorum) {
                         throw new UserException(InternalErrorCode.REPLICA_FEW_ERR,
                                 "Tablet lost replicas. Check if any backend is down or not. tablet_id: "
-                                + tablet.getId() + ", backends: " + Joiner.on(",").join(tablet.getBackends()));
+                                + tablet.getId() + ", backends: " + Joiner.on(",").join(localTablet.getBackends()));
                     }
                     locationParam
                             .addToTablets(new TTabletLocation(tablet.getId(), Lists.newArrayList(bePathsMap.keySet())));
