+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Utils.java
@@ -10,11 +10,11 @@ import com.starrocks.catalog.Column;
 import com.starrocks.catalog.HiveTable;
 import com.starrocks.catalog.HudiTable;
 import com.starrocks.catalog.IcebergTable;
+import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.OlapTable;
 import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.ScalarType;
-import com.starrocks.catalog.Tablet;
 import com.starrocks.catalog.Type;
 import com.starrocks.external.hive.HiveColumnStats;
 import com.starrocks.external.iceberg.cost.IcebergTableStatisticCalculator;
@@ -460,7 +460,7 @@ public class Utils {
             MaterializedIndex materializedIndex = partition.getIndex(selectedIndexId);
             // TODO(kks): improve this for loop
             for (Long id : selectedTabletId) {
-                Tablet tablet = materializedIndex.getTablet(id);
+                LocalTablet tablet = (LocalTablet) materializedIndex.getTablet(id);
                 if (tablet != null && tablet.getQueryableReplicasSize(visibleVersion, schemaHash)
                         != aliveBackendSize) {
                     return false;
