+++ fe/fe-core/src/test/java/com/starrocks/alter/RollupJobV2Test.java
@@ -44,6 +44,7 @@ import com.starrocks.catalog.Database;
 import com.starrocks.catalog.FakeCatalog;
 import com.starrocks.catalog.FakeEditLog;
 import com.starrocks.catalog.KeysType;
+import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.MaterializedIndex.IndexExtState;
 import com.starrocks.catalog.OlapTable;
@@ -229,7 +230,7 @@ public class RollupJobV2Test {
         }
         MaterializedIndex shadowIndex = testPartition.getMaterializedIndices(IndexExtState.SHADOW).get(0);
         for (Tablet shadowTablet : shadowIndex.getTablets()) {
-            for (Replica shadowReplica : shadowTablet.getReplicas()) {
+            for (Replica shadowReplica : ((LocalTablet) shadowTablet).getReplicas()) {
                 shadowReplica.updateRowCount(testPartition.getVisibleVersion(),
                         shadowReplica.getDataSize(),
                         shadowReplica.getRowCount());
@@ -263,7 +264,7 @@ public class RollupJobV2Test {
         assertEquals(Partition.PartitionState.NORMAL, testPartition.getState());
         assertEquals(OlapTableState.ROLLUP, olapTable.getState());
 
-        Tablet baseTablet = baseIndex.getTablets().get(0);
+        LocalTablet baseTablet = (LocalTablet) baseIndex.getTablets().get(0);
         List<Replica> replicas = baseTablet.getReplicas();
         Replica replica1 = replicas.get(0);
         Replica replica2 = replicas.get(1);
@@ -308,7 +309,7 @@ public class RollupJobV2Test {
         }
         MaterializedIndex shadowIndex = testPartition.getMaterializedIndices(IndexExtState.SHADOW).get(0);
         for (Tablet shadowTablet : shadowIndex.getTablets()) {
-            for (Replica shadowReplica : shadowTablet.getReplicas()) {
+            for (Replica shadowReplica : ((LocalTablet) shadowTablet).getReplicas()) {
                 shadowReplica.updateRowCount(testPartition.getVisibleVersion(),
                         shadowReplica.getDataSize(),
                         shadowReplica.getRowCount());
