+++ fe/fe-core/src/test/java/com/starrocks/catalog/AdminStmtTest.java
@@ -95,7 +95,7 @@ public class AdminStmtTest {
         for (Partition partition : tbl.getPartitions()) {
             for (MaterializedIndex index : partition.getMaterializedIndices(IndexExtState.VISIBLE)) {
                 for (Tablet tablet : index.getTablets()) {
-                    for (Replica replica : tablet.getReplicas()) {
+                    for (Replica replica : ((LocalTablet) tablet).getReplicas()) {
                         tabletToBackendList.add(Pair.create(tablet.getId(), replica.getBackendId()));
                     }
                 }
