+++ fe/fe-core/src/test/java/com/starrocks/catalog/LocalTabletTest.java
@@ -36,9 +36,9 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 
-public class TabletTest {
+public class LocalTabletTest {
 
-    private Tablet tablet;
+    private LocalTablet tablet;
     private Replica replica1;
     private Replica replica2;
     private Replica replica3;
@@ -67,7 +67,7 @@ public class TabletTest {
             }
         };
 
-        tablet = new Tablet(1);
+        tablet = new LocalTablet(1);
         TabletMeta tabletMeta = new TabletMeta(10, 20, 30, 40, 1, TStorageMedium.HDD);
         invertedIndex.addTablet(1, tabletMeta);
         replica1 = new Replica(1L, 1L, 100L, 0, 200000L, 3000L, ReplicaState.NORMAL, 0, 0);
@@ -88,10 +88,6 @@ public class TabletTest {
         Assert.assertEquals(replica1, tablet.getReplicaByBackendId(replica1.getBackendId()));
         Assert.assertEquals(replica2, tablet.getReplicaByBackendId(replica2.getBackendId()));
         Assert.assertEquals(replica3, tablet.getReplicaByBackendId(replica3.getBackendId()));
-
-        long newTabletId = 20000;
-        tablet.setTabletId(newTabletId);
-        Assert.assertEquals("tabletId=" + newTabletId, tablet.toString());
     }
 
     @Test
@@ -125,7 +121,7 @@ public class TabletTest {
 
         // 2. Read a object from file
         DataInputStream dis = new DataInputStream(new FileInputStream(file));
-        Tablet rTablet1 = Tablet.read(dis);
+        LocalTablet rTablet1 = LocalTablet.read(dis);
         Assert.assertEquals(1, rTablet1.getId());
         Assert.assertEquals(3, rTablet1.getReplicas().size());
         Assert.assertEquals(rTablet1.getReplicas().get(0).getVersion(), rTablet1.getReplicas().get(1).getVersion());
@@ -134,7 +130,7 @@ public class TabletTest {
         Assert.assertTrue(rTablet1.equals(rTablet1));
         Assert.assertFalse(rTablet1.equals(this));
 
-        Tablet tablet2 = new Tablet(1);
+        LocalTablet tablet2 = new LocalTablet(1);
         Replica replica1 = new Replica(1L, 1L, 100L, 0, 200000L, 3000L, ReplicaState.NORMAL, 0, 0);
         Replica replica2 = new Replica(2L, 2L, 100L, 0, 200000L, 3000L, ReplicaState.NORMAL, 0, 0);
         Replica replica3 = new Replica(3L, 3L, 100L, 0, 200000L, 3000L, ReplicaState.NORMAL, 0, 0);
@@ -144,7 +140,7 @@ public class TabletTest {
         tablet2.addReplica(replica3);
         Assert.assertTrue(tablet2.equals(tablet));
 
-        Tablet tablet3 = new Tablet(1);
+        LocalTablet tablet3 = new LocalTablet(1);
         tablet3.addReplica(replica1);
         tablet3.addReplica(replica2);
         tablet3.addReplica(new Replica(4L, 4L, 100L, 0, 200000L, 3000L, ReplicaState.NORMAL, 0, 0));
